Move to a chip select (CS) mechanism for selecting the memory/io units.
For each memory/io block, add in a chip select (CS) line, simplify the passing of addresses, especially to the video memory bitmap, tilemaps, character map areas.

Update the ps2 controller for the new version of Silice (draft).

ALU try to correctly handle the 32 bit atomic operations, use CS line.

background simplify the writing of the background parameters.

bitmap use CS line for direct memory access.

character_map use CS line for direct memory access.

DMA (io_memmap) eliminate the RGB pixelblock transfer protocol, use to IO with a single register taking R then G then B on subsequent writes.

io_memmap use chip select CS.

MEMORY use chip select lines(CS) for blocks.

PAWS simplify the creation off address + 2, and the lower/upper address0 values, pass CS lines to blocks.

timers_random use chip select (CS) for memory access.

video_memmmap use chip selects CS for registers, bitmaps, tilemaps, character map.

+++ RV64/SOFTWARE/c/library/PAWSlibrary.h
@@ -43,7 +43,7 @@ extern void sleep1khz( unsigned short );
 extern void set_timer1khz( unsigned short, unsigned char );
 extern unsigned short get_timer1khz( unsigned char );
 extern void wait_timer1khz( unsigned char );
-extern unsigned short get_timer1hz( unsigned char );
+extern unsigned long get_timer1hz( unsigned char );
 extern void reset_timer1hz( unsigned char );
 extern unsigned long systemclock( void );
 #ifdef __cplusplus
